Split up logging.cpp/.h into 3 source pairs (1 per struct)
