+++ src/logging/warp_delta_statistics.cpp
@@ -1,9 +1,9 @@
 /*
- * logging.cpp
+ * warp_delta_statistics.cpp
  *
- *  Created on: Nov 13, 2018
+ *  Created on: Mar 15, 2019
  *      Author: Gregory Kramida
- *   Copyright: 2018 Gregory Kramida
+ *   Copyright: 2019 Gregory Kramida
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -19,10 +19,11 @@
  */
 
 //local
-#include "logging.hpp"
+#include "warp_delta_statistics.hpp"
 #include "../math/filtered_statistics.hpp"
 #include "../math/statistics.hpp"
 #include "../math/assessment.hpp"
+
 
 namespace logging {
 WarpDeltaStatistics::WarpDeltaStatistics(
@@ -120,97 +121,4 @@ std::ostream& operator<<(std::ostream &ostr, const WarpDeltaStatistics &ts)
 	return ostr;
 }
 
-TsdfDifferenceStatistics::TsdfDifferenceStatistics(
-		float difference_min,
-		float difference_max,
-		float difference_mean,
-		float difference_standard_deviation,
-		math::Vector2i biggest_difference_location
-		) :
-		difference_min(difference_min),
-				difference_max(difference_max),
-				difference_mean(difference_mean),
-				difference_standard_deviation(difference_standard_deviation),
-				biggest_difference_location(biggest_difference_location)
-{
-}
-
-TsdfDifferenceStatistics::TsdfDifferenceStatistics(const eig::MatrixXf& canonical_field,
-		const eig::MatrixXf& live_field) {
-	eig::ArrayXXf diff = (live_field.array() - canonical_field.array()).abs().eval();
-	float diff_min = diff.minCoeff();
-	eig::ArrayXXf::Index max_row, max_col;
-	float diff_max = diff.maxCoeff(&max_row, &max_col);
-	float diff_mean = diff.mean();
-	float diff_std = std::sqrt((diff - diff_mean).square().sum() / (static_cast<float>(diff.size()) - 1.0f));
-	math::Vector2i diff_max_loc(max_col, max_row);
-
-	this->difference_min = diff_min;
-	this->difference_max = diff_max;
-	this->difference_mean = diff_mean;
-	this->difference_standard_deviation = diff_std;
-	this->biggest_difference_location = diff_max_loc;
-}
-
-eig::VectorXf TsdfDifferenceStatistics::to_array() {
-	eig::VectorXf out(7);
-	out << difference_min,
-			difference_max,
-			difference_mean,
-			difference_standard_deviation,
-			static_cast<float>(biggest_difference_location.x),
-			static_cast<float>(biggest_difference_location.y)
-	;
-	return out;
-}
-
-bool TsdfDifferenceStatistics::operator==(const TsdfDifferenceStatistics& rhs) {
-	return math::almost_equal(this->difference_min, rhs.difference_min) &&
-			math::almost_equal(this->difference_max, rhs.difference_max) &&
-			math::almost_equal(this->difference_mean, rhs.difference_mean) &&
-			math::almost_equal(this->difference_standard_deviation, rhs.difference_standard_deviation) &&
-			this->biggest_difference_location == rhs.biggest_difference_location;
-}
-
-bool TsdfDifferenceStatistics::operator!=(const TsdfDifferenceStatistics& rhs) {
-	return !(*this == rhs);
-}
-
-//TODO: print out one-line JSON representation instead
-std::ostream &operator<<(std::ostream &ostr, const TsdfDifferenceStatistics &ts){
-	ostr << "[tsdf diff stats]"
-			<< std::endl << "  min: " << ts.difference_min
-			<< std::endl << "  max: " << ts.difference_max
-			<< std::endl << "  mean: " << ts.difference_mean
-			<< std::endl << "  std: " << ts.difference_standard_deviation
-			<< std::endl << "  greatest diff at: (" << ts.biggest_difference_location << ")"
-			;
-	return ostr;
-}
-
-ConvergenceReport::ConvergenceReport(int iteration_count,
-		bool iteration_limit_reached,
-		WarpDeltaStatistics warp_delta_statistics,
-		TsdfDifferenceStatistics tsdf_difference_statistics)
-:
-		iteration_count(iteration_count),
-				iteration_limit_reached(iteration_limit_reached),
-				warp_delta_statistics(warp_delta_statistics),
-				tsdf_difference_statistics(tsdf_difference_statistics) {
-}
-
-std::ostream &operator<<(std::ostream &ostr, const ConvergenceReport &ts){
-	ostr << "===[convergence report]==="
-			<< std::endl << "  iter count: " << ts.iteration_count
-			<< std::endl << "  limit reached: " << ts.iteration_limit_reached
-			<< std::endl << "--------------------------"
-			<< std::endl << ts.warp_delta_statistics
-			<< std::endl << "--------------------------"
-			<< std::endl << ts.tsdf_difference_statistics
-			<< std::endl << "==========================)"
-			;
-	return ostr;
-}
-
 } //namespace logging
-
