+++ src/nonrigid_optimization/hierarchical_optimizer2d.hpp
@@ -27,7 +27,7 @@
 
 //local
 #include "../math/tensors.hpp"
-#include "../logging/logging.hpp"
+#include "../logging/convergence_report.hpp"
 
 namespace eig = Eigen;
