+++ src/python_export/logging.cpp
@@ -30,7 +30,9 @@
 
 //local
 #include "logging.hpp"
-#include "../logging/logging.hpp"
+#include "../logging/warp_delta_statistics.hpp"
+#include "../logging/tsdf_difference_statistics.hpp"
+#include "../logging/convergence_report.hpp"
 #include "../math/typedefs.hpp"
 #include "../math/tensors.hpp"
